requirements.txt
Pillow
tqdm
transformers==4.42.4
requests
open_clip_torch==2.26.1
git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
huggingface-hub==0.23.4
numpy
pandas
regex